backend/requirements.txt
numpy>=1.24
scipy>=1.10
trimesh>=4.0
open3d>=0.18
pythonocc-core>=7.8
